ls2084abbmini: added ls2 sdcard factory image
The LS2 sdcard factory (fsl-image-blueboxls2sdfactory.bb) is an image
capable of writing the LS2084a NOR flash of BlueBox Mini board and
update the rcw, ppa, management-complex, rootfilesystem on both banks.
It makes use of the bbdeployscripts to update the flash and uses
fsl-image-auto as the root filesystem.

Issue: ALB-2189
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Vicovan Ionut-Valentin <Ionut.Vicovan@nxp.com>