mesa: Condition removals from PROVIDES/PACKAGECONFIG
Based on the commit "bcc2201454a43c059bcba3851125340d389bd03b"
from the fae/meta-bluebox repository.

Condition removals from PROVIDES/PACKAGECONFIG on whether or not
we are building the image with the Vivante GPU driver. This change
avoids the situation when required functionality is provided by none
if not building the image with said driver.

Issue: ALB-2168
Signed-off-by: Heinz Wrobel <heinz.wrobel@nxp.com>
Signed-off-by: Nica Dan <dan.nica@nxp.com>