t4: Ignore linaro bbappends
Since we do not have meta-linaro for T4 machines, all the bbappend files
under recipes-devtools (mostly gcc) whose version includes "linaro" will
generate errors, as there are no matching recipes. This is not related
to PREFERRED_VERSIONs, but to the versions included in the file names.
In meta-bluebox, we did not have any of these append files.

Issue: ALB-2267
Signed-off-by: Emanuel Muraru <emanuel.muraru@nxp.com>