ubuntu: add basic proxy support
Added basic proxy suport for the pseudo/chroot environment
used when updating the target root filesystem.
This consists of two variables to be defined preferably
in local.conf:
ENV_HOST_PROXIES - a multiline list of proxies, each proxy on a line, e.g.
    ENV_HOST_PROXIES = "http_proxy=http://my.proxy.nxp.com:8080 \
                        https_proxy=http://my.proxy.nxp.com:8080 "
APTGET_HOST_PROXIES - a multiline list of 'Acquire' options to be written to the apt.conf from
                      the target root filesystem, which is used during the filesystem update, e.g.:
    APTGET_HOST_PROXIES = "Acquire::http::proxy \"my.proxy.nxp.com:8080/\"; \
                           Acquire::http::proxy \"my.proxy.nxp.com:8080/\"; "
Normally only the http(s) proxy is required.

Issue: ALB-2402
Signed-off-by: Vicovan Ionut-Valentin <Ionut.Vicovan@nxp.com>