gcc: add dynamic isl support to gcc
ISL 0.16.1 is added as dynamic dependency to gcc builds (native,
cross, target) in a new recipe.
This enables Graphite loop transformation infrastructure and a number
of loop optimization options:
-fgraphite, -fgraphite-identity, -floop-block, -floop-interchange,
-floop-strip-mine, -floop-parallelize-all, -floop-unroll-and-jam,
and -ftree-loop-linear.

In addition, an upstream patch is also cherrypicked, to remove
support for ISL older than 0.15 and fix the incorrect ISL version

hardcoding in GCC.

Issue: ALB-2500
Signed-off-by: Vicovan Ionut-Valentin <Ionut.Vicovan@nxp.com>