images: Re-enable sja1105x and tja110x for all kernel versions
The inclusion of the sja1105x driver into images was disabled
if the kernel version was 4.14, since at that time the sja1105x driver
was not compatible with it. Now, the sja1105x recipe can build a
compatible version of the driver based on the preferred version
for linux. Therefore, the sja1105x should now be included in images
regardless of the linux version.

The tja110x was also disabled, since there was no sja1105x for 4.14
at that point.

Issue: ALB-2450
Signed-off-by: Dan Nica <dan.nica@nxp.com>