linux-s32: run also  "make scripts" after shared work folder is created
Typically, make scripts is executed from the module recipes in task
do_make_scripts
However, a race condition may appear where the modules's make scripts is
executed, then Kernel is cleaned.
On the next module build, make scripts is not executed again therefore
the module failed as the needed Makefiles are not generated.

The fix is to force running "make scripts" for the linux recipe afater
the shared work folder is created.

Issue: ALB-3061

Signed-off-by: Catalin Udma <catalin-dan.udma@nxp.com>