xfce4: downgrade xfce4-terminal to version 0.6.3
Add xfce4-terminal 0.6.3 plus dependencies
(vte 0.28) in order to prevent some
errors and warnings on the startup console
as well as the terminal having an incorect menu appearance.

The downgrade versions of xfce4-terminal and dependencies
are based on gtk+ (2.x).
The version from poky (0.8.6) is based on gtk+3 and has issues
in current environment such as lots of warnings, e.g.
`(xfce4-terminal:1837): Gtk-WARNING **: Theme parsing error: gtk.css`
as well as menu displayed incorrectly (probably due to incompatibilities
with gtk+ 3.22 which is default in rocko).

Issue: ALB-3081
Signed-off-by: Vicovan Ionut-Valentin <Ionut.Vicovan@nxp.com>