fec: limit TX queues to prevent TX starvation crash
Heavy write to NFS (files > 4MB) results in FEC TX queue 0 starvation.
This happens because traffic is shared between the high priority
queues 1 and 2 (used for AVB streaming), which
leads to starvation for low priority queue 0.
This generates exception and dumping of the TX buffers, which
blocks the console and the kernel execution.

Workaround is to use only the default TX queue 0 in order to
prevent starvation.

The workaround is enabled by default for s32v234 by yocto variable
KERNEL_FEC_LIMIT_TX which is set to "1", and can be disabled
(for usecases requiring all TX queues and not involving NFS)
by setting it to "0" in local.conf.

Issue: ALB-2941
Signed-off-by: Vicovan Ionut-Valentin <Ionut.Vicovan@nxp.com>