HOSTTOOLS: Add missing executables to 'HOSTTOOLS' for s32x platforms
In 'do_image_sdcard' task, some executables such as: 'parted',
'mkfs.vfat' and 'mcopy' are required as they are used in sdcard
image creation.
However, those executables are not found during a yocto build
causing an error. To fix this issue, we add the necessary
executables to 'HOSTTOOLS' variable.

Issue: ALB-3326
Signed-off-by: Costea Ciprian <ciprianmarian.costea@nxp.com>