u-boot: Declare 'IMAGE_BOOTLOADER' variable
'IMAGE_BOOTLOADER' is used in burning u-boot to the 'sdcard' image.
It used to be declared in rocko version of 'meta-freescale' layer,
but in sumo version it is not declared anymore. Thus, we declare
'IMAGE_BOOTLOADER' variable in meta-alb layer to ensure that
'u-boot' will be written into the 'sdcard' image.

Issue: ALB-3326
Signed-off-by: Costea Ciprian <ciprianmarian.costea@nxp.com>