ubuntu: Fix package write manifest files check
We want the Yocto recipes as default provider, but we don't
want them to be deployed to avoid pollution of the target rootfs,
since 'ubuntu-base' also provides the packages concerned.
This leads to some package write manifests to not be created which
cause warnings/errors at build time, due to some sanity checks added
in sumo version of poky.
We create those concerned package write manifests to ensure a clean
build of fsl ubuntu images.

Issue: ALB-3326
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>