qemu-user: This patch set fixes the obscure Ubuntu 18 build failures
This patch set contains diagnostic improvements to the strace mechanism of qemu that
were used when debugging the build failures. During the debug, a weakness of
the getsockopt syscall implementation was identified and fixed.
They main fix however is how elf images get loaded. Root cause of the build problems
was loading of images that broke the heap mechanism for the guest. This occurred
semi randomly only because the guest memory allocation is not fully independent of
the host, and the effects were not necessarily obviously related to the problem of
brk() failing at specific times.

Issue: ALB-3523

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>