qemu: SO_PEERGROUPS was not handled correctly, breaking hosts supporting it
If the BSP was built on a host that supports SO_PEERGROUPS, the qemu handling
resulted in a permanent return of ERANGE, leading to a permanent retry and
death of the BSP build with 100% CPU load. This was noticed on a
"getent group bluetooth" command executed in the emulation.
This patch tries to do the right thing about the socket option and thereby
fixes the guest.

Issue: ALB-3523

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>