emu: 11 qemu fixes to support Ubuntu installs
When doing Ubuntu installs, the apt invocation for complex packages
revealed a good number of shortcomings in qemu.
getsockopt had problems.
mmap related stuff needed a number backports and fixes.
To be able to deal with this at all, the qemu-user strace was updated
A real ugly bug in elfload cause spurious and non-reproducable
failures of the emulated code due to breaking host side memory
heap management.

update qemu patches for 2.11.1 version as provided in sumo branch

Issue: ALB-3523

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Catalin Udma <catalin-dan.udma@nxp.com>