ls2084bbminiubuntu: Do not include 'fsl-image-itbflash'
The 'fsl-image-itbflash' image is used as a FALLBACK_RAMDISK in the
'linux-kernelitb' recipe, which in turn is used in the creation of
flash based images.

For '*ubuntu' machines, 'fsl-image-itbflash' fails to build. But since
flash based images are not supported for ubuntu builds, we can use
'linux-kernelitb-norootfs' instead, which does not include 'fsl-image-itbflash'

Issue: ALB-3543
Signed-off-by: Dan Nica <dan.nica@nxp.com>