nativeaptinstall: Renamed/refactored do_shell_update function
THe problem with the function was that a) it was too generically
named despite being very class specific and b) that it couldn't be
deterministically changed. Reason for b) being that appends/prepends
cannot be guaranteed to be in any particular order, so the necessary
base init in the class conflicted with the user side. By renaming
the function into do_aptget_... and splitting the base init and the
user function, the user has a chance to do the right thing without
messing up the class functionality.

ALB-3657
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>