u-boot-qoriq: minor update change where ls2084bbmini update is added
The patches for u-boot-qoriq from  meta-alb will confict with the
patches from meta-vnp, as both are doing updates around the same line
This change is just a minor workaround, moving location where
ls2084bbmini is added just to avoid the conflict with where ls1043adcm
is updated.
Eventually, as other workaround is to apply ls2084abbmini changes only
for ls2084abbmini machine.

The long term approach is to merge the bbmini patches and vnp patches
directly into the u-boot-qoriq git repository.

Issue: ALB-3864

Signed-off-by: Catalin Udma <catalin-dan.udma@nxp.com>