ubuntu: Do not create a flash image for Ubuntu targets
As part of ALB 3478 issue, a flash image will be created for 's32v234evb'
targets. In case of the full Ubuntu target image, build on Ubuntu 18.04
host, it leads to the following error during 'do_image_cpio' task:
cpio: cannot seek on output: Invalid argument
This error is related to the large size of the Ubuntu flash image which
is being created.
Therefore, we temporary remove the flash image creation for every Ubuntu
target images on 's32v234evb' board.

Issue: ALB-3657
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>