ubuntu-ros: Fix 'APTGET_EXTRA_PPA' variable override
Building 'fsl-image-ubuntu-ros' target image on Ubuntu 16.04
host, is failing on 'aptget_update_begin' function due to missing
PPAs. 'APTGET_EXTRA_PPA' is the variable which holds the PPAs
and currently it is overwritten in 'fsl-image-ubuntu-ros.bb'
recipe when the last PPA is added.
We fix this issue by appending the last PPA instead of assigning
the 'APTGET_EXTRA_PPA' variable to it.

Issue: ALB-3657
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>