demos: Fix PCIe demos crash on old boards
The demos crash on S32V234-EVB 28899 and PCIE because the hardware has
insufficient RAM. Those boards have 256 or 512 MB of DDR. The shared
memory buffer is allocated at address 0xFD000000, which is outside the
physical boundary of the memory of these two boards. Move the buffer to
address 0xC1000000.

Issue: ALB-3253
Signed-off-by: Teodor Marina <teodor.marina@nxp.com>