ubuntu-16.04: Disable NetworkManager-wait-online service
NetworkManager-wait-online service is failing on Ubuntu 16.04.
This service runs: '/usr/bin/nm-online -s -q --timeout=30'.

The '-s' parameter changes the default behaviour of the 'nm-online'
executable to not finish successfully when one of the network
interfaces obtained an IPv4 address but to wait for 'NetworkManager'
service startup to complete. Therefore, startup is considered
complete once NetworkManager has activated (or attempted to activate)
every auto-activate connection which is available given the current
network state.

The solution is to mask 'NetworkManager-wait-online' service
as the boot time will increase with 30 seconds if the Ethernet cable
is not connected, even if we remove the '-s' parameter from the
'nm-online' executable.
This delay is not wanted, since either way the network interfaces
are operational if the Ethernet cable is connected before or after
boot reached Ubuntu prompt.

Issue: ALB-4071
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>