sdcard: Dependency handling was confused/incorrect/broken
Due to a misunderstanding and a class design flaw, SDCARD_ROOTFS
has been specified sometimes with DEPLOY_DIR_IMAGE when for
Yocto reasons it would have had to be specified with IMGDEPLOYDIR,
which is an internal variable. This led to a bad prior commit
hardcoding a dependency on ext3 build which was utterly pointless.

The class will now only take the intended rootfs extension and
build all internal names from there. This fixed the build.
if an image build fails, the implementation of Yocto requires
that you -c clean first before trying again. Image builds are
"All or Nothing"!

Given that now only the extension matters, the machine conf
files can now be simplified, eliminating second guesses of the
rootfs file name.

Issue: ALB-4522

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>