sdcard: Unified the image generation function to fix the out-of-sync problem
In a prior commit I created an additional out-of-sync problem in the
image generation functions because I fixed the extra rootfs handling
in the LS function but not in the function used for S32.
To avoid this mistake going forward, there is now a single function
for all. The functions were doing the same things anyway

Issue: ALB-4522

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>