ubuntubasics.inc: Placeholder for proper GLIBC selection
This is in comments because it is not functional yet. The idea
is that we want to build Yocto packages for a "foreign" rootfs
like a Ubuntu one with a glibc that does not exceed the glibc
on the respective Ubuntu version.

Issue: ALB-4611
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
(cherry picked from commit ef861eb00d5b41790991efe43bf418601c279ce7)