Ported glibc 2.23 and 2.27 to Zeus for Ubuntu support.
To be able to properly build Yocto packages that run
in a Ubuntu environment, they need to be built with a glibc
version that does not exceed the Ubuntu glibc version.
We port two versions to match 16.04 and 18.04.
Note that it is not the intent to be 100% option compatible
to Ubuntu. This is about sufficient functionality to make
the basic tools possible that can't be found in a Ubuntu
package repository.
By doing so, there is also no need to use a Ubuntu host or
cross tool chain when the libs are used.

Issue: ALB-4611

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
(cherry picked from commit 9bc819ac6a4f617d3b4211e199c4c941cea4edd7)