mc-utils: meta-alb now relies on the latest from meta-freescale
This is likely easier in maintenance, but also a bit of a complicated
commit. Given that the meta-freescale recipe has a hardcoded board
name "RDB" we need to work around that through the "custom" format
supported in that recipe. We do that by renaming our dpc/dpl-files
appropriately and then relying on the original recipe.

Given that the original recipe is missing family member names like
ls2084a, we have a dynamic-layer addition as that is a generic
problem rather than a meta-alb addition

Issue: ALB-4599

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
(cherry picked from commit d55a71679c96df5ca02d192b1d33fae2ddcfd27f)