nativeaptinstall.bbclass: Intentionally skipping "apt upgrade" now
One problem with upgrading is that the final package list does
not necessarily match our pre-defined list in the ubuntu-base
recipes. This causes weird dependency changes because we can't
adapt on the fly easily right now.
Until we adapt and use BB_INVALIDCONF in magic ways, it is better
to let the customer update on the target. The assumption is that
Ubuntu will always be used on mass storage.

Issue: ALB-4611

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
(cherry picked from commit 1e0dc40629a0678200d7c06d166143f9b9de0613)