ubuntu: preserve dependency order when updating in nodistropackage.inc
The issue was seen as when building kernel-devsrc:
"ERROR: When reparsing
/work/AUTO/yocto/zeus/sources/poky/meta/recipes-kernel/
linux/kernel-devsrc.bb:do_package, the basehash value changed from
    108ec94bf7deaf565a39051ca79586caf227b5893a308650922e446243bcae44 to
        a1146c9283ebadb4286b2df2f1880325056cd0bd45590a4c0ba41f056797a3df.
The metadata is not deterministic and this needs to be fixed."

The data input for computing the hash is different for
RDEPENDS_kernel-devsrc:
the content it's the same but in different order.

The fix envorce the same order for dependency.

Issue: ALB-4639

Signed-off-by: Catalin Udma <catalin-dan.udma@nxp.com>