ubuntu-base: Enabled build of ubuntu-base 20.04
This required better version handling in the global ubuntubasics.inc
because the current version doesn't have a subversion yet. The
general fix here should also help in the future.
We also encode the required glibc properly ... which prompts a
needed version of binutils to build it properly. The assumption
is that we are backwards compatible to 16/18, but that hasn't
beeen build tested yet.

Issue: ALB-4912

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>