u-boot-environment: Fix bloating the result on repeated invocation
While testing the updated Ubuntu code, I stumbled over this
ugly bug. The recipe modified its input data on invocation.
So when it was repeatedly invoked for some reason, it would
repeatedly append the U-Boot default environment, leading
to an ever increasing output file until the environment space
was suddenly exhausted.

This patch changes the method to not touch the original input
file and also to fall back to the last of any duplicated line
in a concatenated environment.

Honestly, I think it is a bad mistake to mix a U-Boot default
environment in here by concatenation, but this needs more
discussion. Given the proper line order and duplicate removal,
any leftover default lines should at least be benign.

Issue: ALB-4912

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>