nativeaptinstall: Fix partial and flaky faketool handling.
/etc/hosts and /etc/resolv.conf were not handled as faketools.
This resulted in bad final output because it was impossible
to determine if they had been replaced. Putting everything
into the unified faketool mechanism and cleaning it up, this
problem is solved. The faketool mechanism needed the cleanup
because corner cases of links, valid or not, were not handled
properly

Issue: ALB-4912

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>