kernel: For iptables.config changed a number of =m to =y
As it turns out, a good number of options is already set in the base
config for the kernel that this builds upon. Whereever the base
config said =y we also say =y now so that do_merge_delta_config does
not report as many changes in the end. We want to align where it
makes sense.

Issue: ALB-5266

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>