sdcard: Eliminated SDCARD_DEPLOYDIR. It couldn't have worked anyway.
The image creation process for a class requires the use of
IMGDEPLOYDIR. Anything else may break the process for dependent
images. The final result of the image creation process ends up
in the user visible DEPLOY_DIR_IMAGE. There is no viable option
to modify this, so SDCARD_DEPLOYDIR was a broken idea.

Issue: ALB-4522

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>