sdcard: The dependency support for the two extra rootfs entries was broken
Anything but the original rootfs of the image itself that is added
into and SD card must come from a different recipe. This includes
the extra root filesystems that are optional. This also means that
there must be a clear dependency check that the other recipes have
been built or we can't pull a rootfs from a different image in!

Issue: ALB-4522

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>