nativeaptinstall: proxy handling was completely broken
First, because of a missing reference to the image root dir, the
proxies were not properly installed at all.
Then the proxies were not parsed in any safe manner. A completely
valid URL could have broken the parsing completely due to the
combination of IFS=" =_".
Finally, the already existing standard host proxies were not
respected even if no class variable for proxies were set.
So in a nutshell, it didn't work right.

Issue: ALB-5218

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>