The rng-tools bbappend broke in a systemd context
This is a workaround, not a fix really. The bbappend we provide
breaks builds completely in systemd context because it was written
systemvinit specific. This workaround enables the systemvinit properly
but does not try to really work any equivalent systemd solution in.

this recipe needs to be revisited anyway because
a) rng-tools exists in a much more recent version in poky
b) It is unclear if this all is relevant with a current kernel

Issue: ALB-5219

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>