image.bbclass needs useradd/groupadd commands
The title isn't quite true. The image.bbclass post install mechanisms
for adding users in systemd need them, which is in
rootfs-postcommands.bbclass used by image.bbclass. If shadow-native
isn't used installed, then these commands will not be there for
building the image and leave errors in the logs that suggest failure.

This patch just ensures that shadow-native is part of the build
process also for Ubuntu based images and eliminates the error messages.

Funny enough, the commands are not really needed because the Ubuntu
files already do the right thing but at least there are no weird
messages now in the log

Issue: ALB-5290

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>