s32-gen1: Add board specific network interfaces file
By default, for s32-gen1 platforms, the 'interfaces' file
which is deployed in the rootfs is taken from the 'poky/meta'
Yocto layer, which is not compatible with our s32-gen1 platforms,
leading to the primary Eth connector to not come up for dhcp, after
Linux boot.

This commit adds board specific network interfaces files
for s32-gen1 platforms.

Issue: ALB-4913
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>