bind:gatesgarth: Enable multithreading
Bind 9.16.7 from gatesgarth doesn't have the option to
disable threads anymore, as it was removed in commit:
[5cdb38c2c7840fad52148bf8668541f4aabba6f7] Remove unthreaded support.

Until now, bind was built with multithreading support disabled,
due to an issue discovered in dhclient, which got stuck in a
pthread_cond_wait() while there is no other thread running.

Bind developers state that POSIX pthread support is fully implemented
in Linux and should be used as such.

Issue: ALB-5918

Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>