nativeaptinstall: Fixing four things, one of them crashing qemu 5
allocation scheme that breaks the brk space. To avoid this problem
we permit access to "/proc/self/maps" to give qemu a chance to find the
right mmap setup
from inside chroot. This was due to the "PWD" variable. We now change
into our chroot dir and patch up PWD to show "/". That fixes the broken
usage in the guest
gatesgarth.

Issue: ALB-6317

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>