bluebox: Converted ls2 variants to use U-Boot 2020.04
This patch updates all the generic U-Boot patches in dynamic layers,
removing most of them effectively.

Building on top of that the BlueBox Mini support is added back in. It
has been completely cleaned up and reworked to fit the new U-Boot.
At this time however, we still run in legacy mode with PPA rather than
converting to ATF (Arm Trusted Firmware, the original name of
TF-A = Trusted Firmware for Armv7-A and Armv8-A).
We are still using ATF rather than TF-A because all recipes used by NXP
in the yocto world are named with 'atf'.

BB Mini now uses the same memory layout that
BB3 uses with the current PPA exception/addition.

To ensure the LS2 images work properly, the mc-utils.git.bbappend had
to be made aware again of the boards to cater to the new recipe format.

A little bit of prep work for using ATF has been added, mostly as
comments in the conf file. A bbappend has been prepared that should
help enable ATF for LS2 in the next step.

The 2017 U-Boot support has been removed.

Issue: ALB-6528

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>