linux: Fix do_rootfs error with DISTRO_FEATURES additions
In case "xen" is added in DISTRO_FEATURES, required configs are
not enabled because SRC_URI is overwritten in the version-specific
linux-s32_*.bb files that include the "linux-s32.inc" file.

The linux-s32.inc file contains the conditional enablements of kernel
configs via config fragments that are appended to SRC_URI, but because
it is overwritten, they are not enabled in the end.

This behavior was introduced by commit
d52e39b560298c654dcafd1ff995314c6e4e747a.

It is fixed by using the common SRC_URI set in the .inc file, while,
for the 5.10 recipe, the temporary branch name is overwritten in the
version-specific .bb file.

Issue: ALB-6736
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>