imagebuilder: Add s32 support patches
Added the patches to support u-boot script generation for s32
platforms, which includes the following:

    - Add "direct-map" and "xen,passthrough" properties for a
DomU's /chosen subnode and the passthroughed device, respectively.
Also, skip following symlinks to the real file name, as this would
add in the u-boot script the real file names from DEPLOY_DIR_IMAGE
which also contain timestamps/machine-specific details.

    - Removed fields from /chosen which are set up by default for
the non-Xen boot flow and break Xen boot (stdout-path,
linux,initrd-start and linux,initrd-end).

    - Add support for prepending path to file names in the final
generated u-boot script, for the use-case where we have the files in a
separate folder that can be accessed with a given $LOAD_CMD.

    - Removed `booti` command from the generated u-boot script,
in order to permit the user to customize the u-boot environment or
the device-tree after executing the script commands before booting.

Issue: ALB-6614, ALB-6618
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>