images: Add fsl-image-dom0less recipe
Added fsl-image-dom0less recipe for generating a .cpio.gz format
rootfs, to be used as an initrd for Dom0less VMs, since they do
not have access to the mmc device.

Dom0less VMs do not have to be Xen-aware, so we do not need any
Xen dependencies installed in their rootfs, therefore, the
rootfs image size is based on core-image-minimal with a few
additions.

Issue: ALB-6618
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>