u-boot: Fix build when both "xen" and "atf" enabled
Xen and ATF are both supported in u-boot by enabling an extra config
at u-boot build time for each of them. However, the name of the
config fragment files that contain the name of the configs are parsed
individually by a yocto task from a list of files, and before this
commit, the names of the files would be concatenated without a
whitespace between them to that list (which is actually a string).

This would generate a build error because of an invalid file name.

This commit fixes it.

Issue: ALB-7091
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>