s32v234: Increase boot size partition
In case of 's32v234evb28899' and 's32v234pcie' platforms,
the sdcard boot size partition was too small (~12MB) to
support Linux Kernel version 5.10, which by itself has a size
of ~12MB.

Solution is to increase it to a default of ~65MB, as it is done
for other S32 platforms.

We also do a small refactoring.

Issue: ALB-7150
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>