sdcard: Add support for extra rootfs partitions with different sizes
Changed the mechanism of creating and writing extra rootfs (EXT) partitions
such that using different images as source for them no longer causes
the build to fail.

Before, in the partition table, each partition would be assumed to
have the same size as the primary one, and trying to write a bigger
partition would cause an error.

Now, we compute the starting offset for each partition before
creating and writing them in the SDCard image, eliminating
the need for the complicated formulas we had before and
the use of PART_NO variable.

Issue: ALB-7429
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>