sysvinit-inittab: xen: Disable getty spawn on default consoles
When running on top of Xen, Linux VMs will use /dev/hvc0 as a console
device, instead of the default console (e.g. /dev/ttyLF0). Moreover,
the default console's device file will not even be created since
Xen will take over the main UART device, and its driver will not
be probed anymore in Linux.

However, in inittab there would still be an entry for respawning
gettys on the default console's device file, and INIT would
continously try to spawn a getty on it, leading to this message
being dumped to console every 5 minutes:
"INIT: Id "LF0" respawning too fast: disabled for 5 minutes"

This commit extends the "inittab" file providing recipe such that
when Xen is enabled via DISTRO_FEATURES, the previously
mentioned lines will be commented, and gettys will not be attempted
to spawn on the default console devices anymore. Commenting over
deleting was chosen to let the user uncomment the line if he
wants to boot native Linux with the same setup.

Issue: ALB-7482
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>