xen-passthrough-dts: Add support for using custom gmac example
The new recipe "xen-examples-dom0less-passthrough-gmac" will use
a different template file as a starting point for the partial DT,
and the "xen-passthrough-dts" recipe will inherit the value of
PASSTHROUGH_NODE variable from an include file specific to the
example set in XEN_EXAMPLE variable, in the case of custom
passthrough examples.

Added support for the custom passthrough example for gmac, and
this can be extended in the future for other similar examples,
using the same approach.

Issue: ALB-7428
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>