xen-examples: Change memory start address
Changed the memory start address such that imagebuilder
will generate the new addresses for loading the required
binaries in RAM in u-boot based on the new address, to
avoid overlapping with reserved memory regions (such as
pfe buffers at 0x83000000-0x84000000).

Otherwise, if we load a binary big enough so it overlaps
with a reserved memory region, the load (and boot) will fail.

Issue: ALB-7614
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>