imagebuilder: Use dynamic sizes for loaded binaries
Imagebuilder would precompute the sizes of the loaded binaries
before generating the script, and the sizes that needed to be
provided to Xen via /chosen would be hardcoded in the boot script.

Switched to using the ${filesize} variable in u-boot, which is
set automatically after a *load command. The value stored by
filesize is now stored in a u-boot env variable with the name
corresponding to the binary that was loaded before. The newly
set variables are now used in setting the /chosen node, instead
of the hardcoded values.

Issue: ALB-7671
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>