image_types_fsl_sdcard: IMAGE_BOOTLOADER may be provided by different recipes
Add a new variable to indicate the recipe that provides U-Boot image to
be deployed to SD card, at it can be different than U-Boot. The
dependency is added against this new variable.

The IMAGE_BOOTLOADER_RECIPE has a default initialization, so no other
recipes updates is needed

Issue: ALB-7700

Signed-off-by: Catalin Udma <catalin-dan.udma@nxp.com>