sysvinit-inittab: Disable getty on ttyLF0 only if Dom0less enabled
Before, for any Xen builds, the /etc/inittab file would be patched
to comment the line that spawned the getty on /dev/ttyLF0 console
device. Since we can use /dev/ttyLF0 as serial console device
in Dom0 (when we are not also booting a Dom0less VM), the line
commenting should only happen if a Dom0less setup is enabled.

Issue: ALB-7888
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>