xen-examples: Add direct-map attribute for basic dom0less VMs
Added the "direct-map" attribute for the Dom0less-DomU booted
when the selected setup is XEN_EXAMPLE = "xen-examples-dom0less".

This is needed for aligning with the other dom0less configurations,
and to avoid printing the virtual memory layout (IPA) of the guest
by Xen at boot time, which can be confusing and leading a user
into thinking those are physical addresses (PA).

Issue: ALB-7999
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>