fsl-image-itbflash: Fix do_package warning
The do_packagedata task from fsl-image-itbflash has to be
removed so that tasks that are relying on it don't print
a warning message in the do_package. In our case
the fsl-image-flash is the one.
This warning is fixed upstream in poky, hardknott branch
commit <b608cf6120235754582edd3e73e34a9af0959b79>.

Issue: ALB-7765
Signed-off-by: Andrei Botila <andrei.botila@nxp.com>