atf: make clean atf build before compile
arm-trusted-firmware can be compiled with different DISTRO_FEATURES, e.g.
optee, m7boot. For each distro feature, there may be different make parameters.
If atf is built then make parameters are changed due to new
DISTRO_FEATURES, the make can not detect the make command line changes.
This result in the elf not being recompiled with the new parameters.

A typical use-case is:
- build ATF without m7boot
- add m7boot in DISTRO_FEATURES
- rebuild fsl-image-* image

ATF is recompiled, the do_compile function is invoke with the new
parameters, but as explained above, the elf is not recompiled.

The issue is fixed be calling a clean before the build.

Issue: ALB-7997

Signed-off-by: Catalin Udma <catalin-dan.udma@nxp.com>