vpp-core: search only in Yocto sysroot for executables
When building lx2160ahpcsom:fsl-image-full while having ccache
installed on the system the vpp-core-native will fail because
it doesn't find ccache at do_compile().
This happends because do_configure() searches for ccache in the
sysroot of Yocto but also in the root of system.
At do_compile() the Yocto sysroot is used which doesn't
contain the ccache so the build fails.
This fix forces ccache to be searched only in the Yocto sysroot.

IssueL ALB-8257
Signed-off-by: Andrei Botila <andrei.botila@nxp.com>