init-ifupdown: Switch PFE port config to static
Changed the default port configuration for PFE to be static,
not by DHCP, due to problems occuring when connecting the
board to a network that has no DHCP server.

Requesting an IP from a DHCP server (if it exists) can be done
manually, for example, by running: `udhcpc -i pfe0`.

Issue: ALB-8766
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>