fsl-image-flash: Fix breakage of Yocto rules for IMAGE_FSTYPES
The Yocto docs clearly state that IMAGE_FSTYPES must be set prior
to inheriting the image class ... which we do when creating the
empty rootfs. We can also depend now on Yocto dependency
management, so there doesn't seem a need to do a convoluted stunt
with IMAGE_TYPES_MASKED anymore.

Issue: ALB-8857
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>