ubuntu: Remove flashimage from ubuntu images due to size constraints
Previously, not building a .flashimage was spread in the individual
machine config files. Given even the smallest size of a reasonable
Ubuntu rootfs, it is however generally not feasible to have a classic
boot flash with a Ubuntu rootfs. This means that we can eliminate
all individual snippets and simply disallow the .flashimage in
ubuntubasics.inc

Issue: ALB-8857
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>