imagebuilder: Add support for appending extra commands to script
Added an imagebuilder patch which enables the tool to add some
user-specified commands to the generated script as "fixups",
before the boot command.

The commands need to be written in a text file, which is then
passed to imagebuilder via "-a <path_to_file>" parameter.

Also removed an old patch which hardcoded a command inside
imagebuilder's sources.

Issue: ALB-8633
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>