xen-examples: Use an EXTRA_CMDS_FILE for appending commands to boot.scr
Added support for using a file mentioned in "EXTRA_CMDS_FILE" variable
to add extra commands as fixups in the boot script generated when booting
with Xen enabled.

The old "fdt rm /chosen stdout-path" command which was hardcoded inside
Imagebuilder's sources is now added to the script via the mentioned file.

Issue: ALB-8633
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>