linux-s32: xen: Enable CONFIG_XEN_GNTDEV as built-in
If "xen" is enabled, the CONFIG_XEN_GNTDEV needs to be set
to "y" instead of "m", such that the gntdev.c driver is
built-in with the kernel image, instead of deployed as an
external module.

This fixes a dependency issue for SystemD modules loading service
which would not find the gntdev object and would prevent the other
xen daemons/services to start.

Issue: ALB-9073
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>