optee: add support for tee-supplicant plugins
OP-TEE has introduced tee-supplicant plugins in a more recent version,
making it possible to add new features to the tee-supplicant in a
fashion that would not alter the upstream version of the supplicant.

Compile the plugins of the optee-examples and optee-test repositories
and place them at the default path expected by the supplicant:
'/usr/lib/tee-supplicant/plugins/'.

Issue: ALB-9320
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>